correct write behavior to leverage sequence order
"leftovers" were being added back in the list of sequences to write,
causing them to have to iterate through the slices list again. Instead
we should continue to write the leftovers since they'll be in the next
slice down.

Also fix cases of multiple identical creates happening that was called
out in some of the new tests